+++ src/uni_modules/uview-plus/components/u-box/box.js
@@ -1,0 +1,24 @@
+/*
+ * @Author       : ly
+ * @Description  :
+ * @version      : 4.0
+ * @Date         : 2026-02-02 23:54:58
+ * @LastAuthor   : ly
+ * @lastTime     : 2026-02-02 23:54:58
+ * @FilePath     : /uview-ultra/components/up-box/box.js
+ */
+export default {
+    // box 组件
+    box: {
+        bgColors: ['#EEFCFF', '#FCF8FF', '#FDF8F2'],
+        height: '160px',
+        borderRadius: '6px',
+        gap: '15px',
+        leftIcon: '',
+        leftTitle: '左',
+        rightTopIcon: '',
+        rightTopTitle: '右上',
+        rightBottomIcon: '',
+        rightBottomTitle: '右下'
+    }
+}